Crate eui

Tutorial

Step 1: build a hierarchy of objects representing the state of the UI.

Create a structure of objects that represent the state of the user interface. Use Arc throughout the hierarchy.

let main_menu = MainMenuUiState {
    title: Arc::new(TextLabelUi { text: "My awesome game!" }),
    buttons: vec![
        Arc::new(ButtonState { title: "Start" }),
        Arc::new(ButtonState { title: "Options" }),
        Arc::new(ButtonState { title: "Quit" }),
    ]
}

Step 2: implement the Widget trait on your ui elements.

This trait implementation tells eui what the layout of your user interface is.

impl eui::Widget for MainMenuUiState {
    fn build_layout(&self) -> eui::Layout {
        let mut list = vec![self.title.clone()];
        for b in &self.buttons { list.push(b.clone()); }
        eui::Layout::Vertical(list)
    }
}

Step 3: create a eui::Ui object and pass your state to it.

The eui library will build a secondary hierarchy that contains the positions, rotation and scale of each component. This hierarchy is handled internally.

At any moment you can call the draw method on your Ui to obtain a list of things to draw when you render your user interface.

You can still access and/or modify your state through the widget() method of the Ui object.

Step 4: send events to the eui::Ui

Whenever the mouse cursor is moved, a key is pressed, etc. you should call the appropriate method on the Ui object to update it.

These calls will be propagated to the methods of the Widget trait so that the state can update itself.

Modules

predefined

Structs

Alignment

Child

EventOutcome
Structure returned by handle_event, indicating information back to the library.

Matrix
A 3x3 matrix. The data is stored in column-major.

Ui
Main struct of this library. Handles the UI as a whole.

Enums

HorizontalAlignment

Layout

Shape
A shape that can be drawn by any of the UI's components.

VerticalAlignment

Traits

Widget
Implement this on types that contain the state of a widget.