Enum eui::Layout

pub enum Layout {
    AbsolutePositionned(Vec<(Matrix, Arc<Widget>)>),
    HorizontalBar {
        alignment: HorizontalAlignment,
        children: Vec<Child>,
        vertical_align: bool,
    },
    VerticalBar {
        alignment: VerticalAlignment,
        children: Vec<Child>,
        horizontal_align: bool,
    },
    Shapes(Vec<Shape>),
}

Variants

AbsolutePositionned

HorizontalBar

The content of the widget will be split in parts whose size depend on the weight of each child. Then the widgets are moved so that the white spaces at the left and right of each child whose collapse value is true are merged with the neighbouring widgets. The alignment is taken into account in order to align the elements once the children have been collapsed.

Fields

alignment

How the children should be aligned once white spaces have been collapsed.

children

List of children.

vertical_align

VerticalBar

The same as HorizontalBar, but vertical.

Fields

alignment

How the children should be aligned once white spaces have been collapsed.

children

List of children.

horizontal_align

Shapes