Enum eui::Shape

pub enum Shape {
    Text {
        matrix: Matrix,
        text: String,
    },
    Image {
        matrix: Matrix,
        name: String,
    },
}

A shape that can be drawn by any of the UI's components.

The meaning of the matrix depends on the context in which the shape is manipulated. When returned by build_layout, the matrix is relative to the widget. When returned by draw, the matrix is absolute (ie. relative to the viewport).

Variants

Text
Fields

matrix

text

Image
Fields

matrix

name

Methods

impl Shape

fn apply_matrix(self, outer: &Matrix) -> Shape

fn get_bounding_box(&self) -> (f32, f32, f32, f32)

Returns the bounding box of the shape.

Order: top, right, bottom, left. Each value is between -1.0 and 1.0. The top will always be greater than the bottom. The right will always be greater than the left.

fn hit_test(&self, point: &[f32; 2]) -> bool

Returns true if the point's coordinates hit the shape.