Struct eui::Ui

pub struct Ui<S> {
    // some fields omitted
}

Main struct of this library. Handles the UI as a whole.

Methods

impl<S> Ui<S> where S: Widget

fn new(state: S, viewport_height_per_width: f32) -> Ui<S>

Builds a new Ui.

fn rebuild(&self)

Rebuilds the UI after the state has been changed.

fn draw(&self) -> Vec<Shape>

"Draws" the UI by returning a list of shapes. The list is ordered from bottom to top (in other words, shapes at the start of the list can be obstructed by shapes further ahead in the list).

The matrices stored in the shapes assume that the viewport uses OpenGL coordinates. This means that the viewport has a width of 2, a height of 2, and that the origin is at the center of the screen.

fn set_viewport_height_per_width(&self, value: f32)

Changes the height per width ratio of the viewport and rebuilds the UI.

fn set_cursor(&self, cursor: Option<[f32; 2]>, down: bool)

Sets the position and state of the cursor.

This function will search for shapes that collide with the cursor and send mouse events to their owner.

fn is_hovering(&self) -> bool

Returns true if the mouse is hovering one of the elements of the UI.

fn widget(&self) -> &S

Returns a reference to the main widget stored in the object.

Note that the UI won't be rebuilt after calling this function. You have to manually call the rebuild() method.