Trait eui::Widget

pub trait Widget: Send + Sync + 'static {
    fn build_layout(&self, height_per_width: f32, alignment: Alignment) -> Layout;

    fn needs_rebuild(&self) -> bool { ... }
    fn handle_event(&self, _event: &Any, _source_child: Option<usize>) -> EventOutcome { ... }
}

Implement this on types that contain the state of a widget.

Required Methods

fn build_layout(&self, height_per_width: f32, alignment: Alignment) -> Layout

Returns a structure indicating the content of this widget.

The height_per_width contains the ratio of the height of the widget divided by its width. The alignment is just an indication passed by the parent.

Provided Methods

fn needs_rebuild(&self) -> bool

This method is called before drawing. It should return true if the layout of this element should be rebuilt.

The default implementation just returns false.

This method is useful when dealing with animation. As long as the widget's content is moving, you should return true.

fn handle_event(&self, _event: &Any, _source_child: Option<usize>) -> EventOutcome

The widget received an event. It can update itself, then it should return an EventOutcome indicating the library what to do next. The default implementation returns Default::default().

The event can be:

predefined::MouseEnterEvent

predefined::MouseLeaveEvent

predefined::MouseClick

Any other event produced by another widget.